pyproject.toml
[project]
name = "nnpid-sitl-mujoco"
version = "0.1.0"
description = "SITL + MuJoCo Drone Simulation Framework with Gymnasium RL Support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "NNPID Team"}
]
keywords = ["drone", "simulation", "mujoco", "px4", "sitl", "reinforcement-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mujoco>=3.0",
    "pymavlink>=2.4",
    "gymnasium>=0.29",
    "numpy>=1.24",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "stable-baselines3>=2.0",
    "tensorboard>=2.14",
    "matplotlib>=3.7",
    "imageio>=2.31",
    "torch>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
nnpid-sitl = "scripts.run_sitl_sim:main"
nnpid-train = "scripts.train_controller:main"
nnpid-eval = "scripts.evaluate:main"
nnpid-viz = "scripts.visualize:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
